pyproject.toml
[tool.poetry]
name = "ocr_project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.1"
pillow = "^9.0.0"
pytesseract = "^0.3.9"
redis = "^5.0.0"
python-multipart = "^0.0.6"
validators = "^0.18.2" 
uvicorn = "^0.23.2"
httpx = "^0.21.1"
dynaconf = "^3.2.0"
phonenumbers = "^8.12.29"
chardet = "^4.0.0"
dateparser = "^1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
